+++ src/main.ts
@@ -37,6 +37,9 @@ import {
 import { OverrideModal } from "./model/override";
 import { path } from "./lib/path";
 import { deduplicateNewName } from "./lib/deduplicate";
+import CreateProcessor from "./create";
+import RenameProcessor from "./rename";
+import { getActiveFile } from "./commons";
 
 export default class AttachmentManagementPlugin extends Plugin {
   settings: AttachmentManagementPluginSettings;
@@ -64,7 +67,7 @@ export default class AttachmentManagementPlugin extends Plugin {
       id: "override-setting",
       name: "Overriding Setting",
       checkCallback: (checking: boolean) => {
-        const file = this.getActiveFile();
+        const file = getActiveFile(this.app);
         if (file) {
           if (!checking) {
             const { setting } = getOverrideSetting(this.settings, file);
@@ -81,7 +84,7 @@ export default class AttachmentManagementPlugin extends Plugin {
       id: "reset-override-setting",
       name: "Reset Override Setting",
       checkCallback: (checking: boolean) => {
-        const file = this.getActiveFile();
+        const file = getActiveFile(this.app);
         if (file) {
           if (!checking) {
             delete this.settings.overridePath[file.path];
@@ -129,8 +132,10 @@ export default class AttachmentManagementPlugin extends Plugin {
           if (isMarkdownFile(file.extension) || isCanvasFile(file.extension)) {
             return;
           }
+
+          const processor = new CreateProcessor(this.app, this.settings);
           if (isImage(file.extension) || isPastedImage(file)) {
-            this.processAttach(file);
+            processor.processAttach(file);
           } else {
             if (this.settings.handleAll) {
               debugLog("create - handleAll for file", file);
@@ -138,7 +143,7 @@ export default class AttachmentManagementPlugin extends Plugin {
                 debugLog("create - excluded file by extension", file);
                 return;
               }
-              this.processAttach(file);
+              processor.processAttach(file);
             }
           }
         });
@@ -190,7 +195,8 @@ export default class AttachmentManagementPlugin extends Plugin {
             debugLog("rename - RENAME_EVENT_TYPE:", RENAME_EVENT_TYPE_FILE);
           }
 
-          await this.onRename(file, oldPath, eventType, type, setting);
+          const processor = new RenameProcessor(this.app, this.settings);
+          await processor.onRename(file, oldPath, eventType, type, setting);
         } else if (file instanceof TFolder) {
           // ignore rename event of folder
           // debugLog("rename - ignore rename folder event:", file.name, oldPath);
@@ -276,336 +282,6 @@ export default class AttachmentManagementPlugin extends Plugin {
   //     }
   //   }
   // }
-
-  async onRename(
-    file: TAbstractFile,
-    oldPath: string,
-    eventType: RenameEventType,
-    attachRenameType: ATTACHMENT_RENAME_TYPE = ATTACHMENT_RENAME_TYPE.NULL,
-    setting: AttachmentPathSettings
-  ) {
-    const rf = file as TFile;
-
-    // old note path and name
-    const oldNotePath = path.dirname(oldPath);
-    const oldNoteName = path.basename(oldPath, path.extname(oldPath));
-    // parent of oldNotePath
-    const oldNoteParent = path.basename(path.dirname(oldPath));
-
-    debugLog("onRename - old note path:", oldNotePath);
-    debugLog("onRename - old note name:", oldNoteName);
-
-    const { parentPath, parentName } = getParentFolder(rf);
-    // old attachment path
-    const oldAttachPath = this.getAttachmentPath(oldNoteName, oldNotePath, oldNoteParent, setting);
-    const newAttachPath = this.getAttachmentPath(rf.basename, parentPath, parentName, setting);
-
-    debugLog("onRename - old attachment path:", oldAttachPath);
-    debugLog("onRename - new attachment path:", newAttachPath);
-
-    // if the old attachment folder does not exist, skip
-    // this will happen when we have already rename the attachment file or folder in previous rename event
-    if (!(await this.app.vault.adapter.exists(oldAttachPath))) {
-      debugLog("onRename - attachment path does not exist:", oldAttachPath);
-      return;
-    }
-
-    if (!(await this.app.vault.adapter.exists(newAttachPath))) {
-      await this.app.vault.adapter.mkdir(newAttachPath);
-    }
-
-    let oldName = "";
-    let newName = "";
-    if (
-      eventType === RENAME_EVENT_TYPE_FILE &&
-      (attachRenameType === ATTACHMENT_RENAME_TYPE.FILE || attachRenameType === ATTACHMENT_RENAME_TYPE.BOTH)
-    ) {
-      oldName = oldNoteName;
-      newName = rf.basename;
-    }
-
-    // rename attachment folder first
-    await this.renameFolder(oldAttachPath, newAttachPath, oldName, newName, attachRenameType);
-
-    // rename attachment filename as needed
-    if (
-      eventType === RENAME_EVENT_TYPE_FILE &&
-      (attachRenameType === ATTACHMENT_RENAME_TYPE.FILE || attachRenameType === ATTACHMENT_RENAME_TYPE.BOTH)
-    ) {
-      // suppose the attachment folder already renamed
-      await this.renameFiles(oldAttachPath, newAttachPath, false, oldName, newName);
-    }
-  }
-
-  async renameFolder(
-    oldAttachPath: string,
-    newAttachPath: string,
-    oldName: string,
-    newName: string,
-    attachRenameType: ATTACHMENT_RENAME_TYPE
-  ) {
-    const { stripedSrc, stripedDst } = stripPaths(oldAttachPath, newAttachPath);
-
-    debugLog("onRename - striped source:", stripedSrc);
-    debugLog("onRename - striped destination:", stripedDst);
-
-    if (stripedSrc === stripedDst) {
-      debugLog("onRename - same striped path");
-    }
-
-    if (
-      stripedSrc !== stripedDst &&
-      (attachRenameType === ATTACHMENT_RENAME_TYPE.FOLDER || attachRenameType === ATTACHMENT_RENAME_TYPE.BOTH)
-    ) {
-      const exitsDst = await this.app.vault.adapter.exists(stripedDst);
-      if (exitsDst) {
-        debugLog("renameFolder - target folder exists:", stripedDst);
-        // move the files in oldAttachPath to newAttachPath
-        await this.renameFiles(oldAttachPath, newAttachPath, true, oldName, newName);
-        // rm the old folder if it's empty
-        const old = await this.app.vault.adapter.list(oldAttachPath);
-        if (old.files.length === 0 && old.folders.length === 0) {
-          await this.app.vault.adapter.rmdir(oldAttachPath, false);
-        }
-        return;
-      } else {
-        const src = this.app.vault.getAbstractFileByPath(oldAttachPath);
-        if (src === null) {
-          debugLog("renameFolder - source file not exists:", oldAttachPath);
-          return;
-        }
-        await this.app.fileManager.renameFile(src, newAttachPath);
-      }
-    }
-  }
-
-  async renameFiles(srcPath: string, dstPath: string, exists: boolean, oldName: string, newName: string) {
-    let attachmentFiles: ListedFiles;
-    if (exists) {
-      attachmentFiles = await this.app.vault.adapter.list(srcPath);
-    } else {
-      attachmentFiles = await this.app.vault.adapter.list(dstPath);
-    }
-
-    debugLog("renameFiles - attachmentFiles:", attachmentFiles);
-
-    // rename all attachment files that the filename content the ${notename} in attachment path
-    for (const filePath of attachmentFiles.files) {
-      let fileName = path.basename(filePath);
-      const fileExtension = path.extname(fileName);
-      if (
-        (this.settings.handleAll && testExcludeExtension(fileExtension, this.settings.excludeExtensionPattern)) ||
-        !isImage(fileExtension)
-      ) {
-        debugLog("renameFiles - no handle extension:", fileExtension);
-        continue;
-      }
-
-      fileName = fileName.replace(oldName, newName);
-
-      const dstFolder = this.app.vault.getAbstractFileByPath(dstPath) as TFolder,
-        { name } = await deduplicateNewName(fileName, dstFolder),
-        newFilePath = normalizePath(path.join(dstPath, name)),
-        src = this.app.vault.getAbstractFileByPath(filePath);
-
-      debugLog("renameFiles - new file path:", newFilePath);
-      if (src === null) {
-        continue;
-      }
-      await this.app.fileManager.renameFile(src, newFilePath);
-    }
-  }
-
-  /**
-   * Post-processing of created attachment file (for paste and drop event).
-   * @param file - the file to process
-   * @returns - none
-   */
-  async processAttach(file: TFile) {
-    const activeFile = this.getActiveFile();
-    if (activeFile === undefined) {
-      new Notice("Error: no active file found.");
-      return;
-    }
-    const { setting } = getOverrideSetting(this.settings, activeFile);
-
-    debugLog("processAttach - active file path", activeFile.path);
-
-    const { parentPath, parentName } = getParentFolder(activeFile);
-
-    debugLog("processAttach - parent path:", parentPath);
-
-    const attachName = this.getPastedImageFileName(activeFile.basename, file.basename, setting) + "." + file.extension;
-    const attachPath = this.getAttachmentPath(activeFile.basename, parentPath, parentName, setting);
-
-    // make sure the path was created
-    if (!(await this.app.vault.adapter.exists(attachPath))) {
-      await this.app.vault.adapter.mkdir(attachPath);
-    }
-
-    const attachPathFile = this.app.vault.getAbstractFileByPath(attachPath) as TFolder;
-    const { name } = await deduplicateNewName(attachName, attachPathFile);
-
-    debugLog("processAttach - new path of file:", path.join(attachPath, name));
-
-    await this.renameCreateFile(file, attachPath, name, activeFile, true);
-  }
-
-  /**
-   * Rename the file specified by `@param file`, and update the link of the file if specified updateLink
-   * @param file - file to rename
-   * @param attachPath - where to the renamed file will be move to
-   * @param attachName - name of the renamed file
-   * @param activeFile - associated active file
-   * @param updateLink - whether to replace the link of renamed file
-   * @returns - none
-   */
-  async renameCreateFile(file: TFile, attachPath: string, attachName: string, activeFile: TFile, updateLink?: boolean) {
-    debugLog("renameFile - src of rename:", file.path);
-    const dst = normalizePath(path.join(attachPath, attachName));
-    debugLog("renameFile - dst of rename:", dst);
-
-    const oldLinkText = this.app.fileManager.generateMarkdownLink(file, activeFile.path);
-    const oldPath = file.path;
-    const oldName = file.name;
-
-    try {
-      // this api will not update the link automatically on `create` event
-      await this.app.fileManager.renameFile(file, dst);
-      new Notice(`Renamed ${oldName} to ${attachName}.`);
-    } catch (err) {
-      new Notice(`Failed to rename ${file.path} to ${dst}`);
-      throw err;
-    }
-
-    if (!updateLink) {
-      return;
-    }
-
-    // in case fileManager.renameFile may not update the internal link in the active file,
-    // we manually replace the current line by manipulating the editor
-    const newLinkText = this.app.fileManager.generateMarkdownLink(file, activeFile.path);
-    debugLog("renameFile - replace text:", oldLinkText, newLinkText);
-
-    const view = this.getActiveView();
-    if (view === null) {
-      new Notice(`Failed to update link in ${activeFile.path}: no active view`);
-      return;
-    }
-    const content = view.getViewData();
-    let val = "";
-    switch (activeFile.extension) {
-      case "md":
-        val = content.replace(oldLinkText, newLinkText);
-        break;
-      case "canvas":
-        val = content.replace(`/(file\s*\:\s*\")${oldPath}(\")/g`, `$1${dst}$2`);
-        break;
-    }
-
-    view.setViewData(val, false);
-    new Notice(`Updated 1 link in ${activeFile.path}`);
-  }
-
-  /**
-   * Return the active text file, `md` or `canvas`
-   * @returns - the active file or undefined if no active file
-   */
-  getActiveFile(): TFile | undefined {
-    const view = this.getActiveView();
-    return view?.file;
-  }
-
-  /**
-   * Return the active view of text file
-   * @returns - the active view of text file
-   */
-  getActiveView(): TextFileView | null {
-    return this.app.workspace.getActiveViewOfType(TextFileView);
-  }
-
-  /**
-   * Generate the attachment path with specified variables
-   * @param noteName - basename (without extension) of note
-   * @param notePath - path of note
-   * @param setting
-   * @param parentFolderBasename
-   * @returns attachment path
-   */
-  getAttachmentPath(
-    noteName: string,
-    notePath: string,
-    parentFolderBasename: string,
-    setting: AttachmentPathSettings = this.settings.attachPath
-  ): string {
-    const root = this.getRootPath(notePath, setting);
-
-    const attachPath = path.join(
-      root,
-      setting.attachmentPath
-        .replace(`${SETTINGS_VARIABLES_NOTEPATH}`, notePath)
-        .replace(`${SETTINGS_VARIABLES_NOTENAME}`, noteName)
-        .replace(`${SETTINGS_VARIABLES_NOTEPARENT}`, parentFolderBasename)
-    );
-    return normalizePath(attachPath);
-  }
-
-  /**
-   * Get root path to save attachment file
-   * @param notePath - path of note
-   * @param setting
-   * @returns root path to save attachment file
-   */
-  getRootPath(notePath: string, setting: AttachmentPathSettings = this.settings.attachPath): string {
-    let root: string;
-
-    //@ts-ignore
-    const obsmediadir = app.vault.getConfig("attachmentFolderPath");
-    // debugLog("obsmediadir", obsmediadir);
-    switch (setting.saveAttE) {
-      case `${SETTINGS_ROOT_INFOLDER}`:
-        root = path.join(setting.attachmentRoot);
-        break;
-      case `${SETTINGS_ROOT_NEXTTONOTE}`:
-        root = path.join(notePath, setting.attachmentRoot.replace("./", ""));
-        break;
-      default:
-        if (obsmediadir === "/") {
-          // in vault root folder case
-          root = obsmediadir;
-        } else if (obsmediadir === "./") {
-          // in current folder case
-          root = path.join(notePath);
-        } else if (obsmediadir.match(/\.\/.+/g) !== null) {
-          // in subfolder case
-          root = path.join(notePath, obsmediadir.replace("./", ""));
-        } else {
-          // in specified folder case
-          root = obsmediadir;
-        }
-    }
-
-    return root === "/" ? root : normalizePath(root);
-  }
-
-  /**
-   * Generate the image file name with specified variable
-   * @param noteName - basename (without extension) of note
-   * @param originalName - original name of attachment file
-   * @param setting
-   * @returns image file name
-   */
-  getPastedImageFileName(
-    noteName: string,
-    originalName: string,
-    setting: AttachmentPathSettings = this.settings.attachPath
-  ): string {
-    const dateTime = window.moment().format(this.settings.dateFormat);
-    return setting.attachFormat
-      .replace(`${SETTINGS_VARIABLES_DATES}`, dateTime)
-      .replace(`${SETTINGS_VARIABLES_NOTENAME}`, noteName)
-      .replace(`${SETTINGS_VARIABLES_ORIGINALNAME}`, originalName);
-  }
 
   backupConfigs() {
     //@ts-ignore
